added merge-coverage-report hook
