+++ .pre-commit-hooks.yaml
@@ -64,3 +64,10 @@
   language: python
   pass_filenames: false
   args: ["-i", ".coverage"]
+
+- id: merge-coverage-report
+  name: merge-coverage-report
+  entry: merge-coverage-report
+  language: python
+  require_serial: true
+  pass_filenames: false
